pyproject.toml
[project]
name = "mistral-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-mistralai>=1.1.1",
    "langfuse>=3.14.5",
    "pydantic-settings>=2.13.1",
    "mistralai>=1.0.0",
    "sqlalchemy>=2.0.47",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.18",
    "scikit-learn>=1.8.0",
]
